off_t in clip2
Former-commit-id: 420648477b4d8fe2c207e8c89250a1265b143f67

+++ user/jennings/Mclip2.c
@@ -36,7 +36,7 @@ sfclip < in.rsf > out.rsf clip=x
 
 int main(int argc, char* argv[])
 {
-    int i, n, nbuf;
+    off_t i, n, nbuf;
     float upper, lower, *trace=NULL;
     sf_file in=NULL, out=NULL; /* Input and output files */
